(refactor): move test-utils/ test/utils/
- going to move tests there too in order to declutter app root
