+++ browserify.config.js
@@ -2,7 +2,7 @@ const browserify = require('browserify')
 const fs = require('fs')
 
 browserify({
-  entries: './test-utils/browserify/_export-to-window.js',
+  entries: './test/utils/browserify/_export-to-window.js',
   // don't parse ammo
   noParse: ['vendor/ammo.js'],
   // don't bundle three
