--- test-utils/browserify/_export-to-window.js
@@ -1,3 +1,0 @@
-var Physijs = require('../../browserify.js');
-
-window.Physijs = Physijs;
