--- test-utils/webpack/_export-to-window.js
@@ -1,3 +1,0 @@
-var Physijs = require('../../webpack.js');
-
-window.Physijs = Physijs;
