--- test-utils/webpack/_import-bundle-to-module.js
@@ -1,3 +1,0 @@
-require('../../build/webpack.bundle.js');
-
-module.exports = window.Physijs;
