+++ test/utils/browserify/_export-to-window.js
@@ -1,0 +1,3 @@
+var Physijs = require('../../../browserify.js');
+
+window.Physijs = Physijs;
