+++ test/utils/browserify/_import-bundle-to-module.js
@@ -1,0 +1,3 @@
+require('../../../build/browserify.bundle.js');
+
+module.exports = window.Physijs;
