+++ test/utils/webpack/_export-to-window.js
@@ -1,0 +1,3 @@
+var Physijs = require('../../../webpack.js');
+
+window.Physijs = Physijs;
