+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path')
 
 module.exports = {
-  entry: './test-utils/webpack/_export-to-window.js',
+  entry: './test/utils/webpack/_export-to-window.js',
   output: {
     path: path.join(__dirname, '/build/'),
     filename: 'webpack.bundle.js',
